modify: WritingFormatterをインターフェースとし,ビルダクラスを独立
